* Objective
This commit performs some UI improvements, especially adding a CTA on the
Home to access the Docs.

* Why

The Home still needs improvement, but adding the CTA was necessary for
proper navigation between the Home and the Docs.

* How

This commit includes the CTA on the Home and performs some minor style
refactors.